01_basic_agent/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "debugpy>=1.8.20",
    "fastapi[standard]>=0.135.1",
    "langchain-core>=1.2.17",
    "langgraph>=1.0.10",
    "litellm>=1.82.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
